+++ .github/actions/publish-libs/npm-publish.sh
@@ -12,7 +12,6 @@ fi
 
 export PROJECTS=(
   'aca-content'
-  'aca-preview'
   'aca-shared'
 );
