+++ .github/actions/update-library-versions/update-lib-versions.sh
@@ -11,7 +11,6 @@ fi
 
 export PROJECTS=(
   'aca-content'
-  'aca-preview'
   'aca-shared'
 );
