+++ projects/aca-content/src/lib/aca-content.routes.ts
@@ -120,7 +120,7 @@ export const CONTENT_LAYOUT_ROUTES: Route = {
         // deprecated, backwards compatibility with ACA 1.8
         {
           path: 'preview/:nodeId',
-          loadChildren: () => import('@alfresco/aca-preview').then((m) => m.PreviewModule),
+          loadChildren: () => import('@alfresco/aca-content/preview').then((m) => m.PreviewModule),
           data: {
             navigateSource: 'personal-files'
           }
@@ -180,7 +180,7 @@ export const CONTENT_LAYOUT_ROUTES: Route = {
         // deprecated, backwards compatibility with ACA 1.8
         {
           path: 'preview/:nodeId',
-          loadChildren: () => import('@alfresco/aca-preview').then((m) => m.PreviewModule),
+          loadChildren: () => import('@alfresco/aca-content/preview').then((m) => m.PreviewModule),
           data: {
             navigateSource: 'personal-files'
           }
@@ -188,7 +188,7 @@ export const CONTENT_LAYOUT_ROUTES: Route = {
         // deprecated, backwards compatibility with ACA 1.8
         {
           path: ':folderId/preview/:nodeId',
-          loadChildren: () => import('@alfresco/aca-preview').then((m) => m.PreviewModule),
+          loadChildren: () => import('@alfresco/aca-content/preview').then((m) => m.PreviewModule),
           data: {
             navigateSource: 'personal-files'
           }
@@ -235,7 +235,7 @@ export const CONTENT_LAYOUT_ROUTES: Route = {
         // deprecated, backwards compatibility with ACA 1.8
         {
           path: 'preview/:nodeId',
-          loadChildren: () => import('@alfresco/aca-preview').then((m) => m.PreviewModule),
+          loadChildren: () => import('@alfresco/aca-content/preview').then((m) => m.PreviewModule),
           data: {
             navigateSource: 'libraries'
           }
@@ -347,7 +347,7 @@ export const CONTENT_LAYOUT_ROUTES: Route = {
         // deprecated, backwards compatibility with ACA 1.8
         {
           path: 'preview/:nodeId',
-          loadChildren: () => import('@alfresco/aca-preview').then((m) => m.PreviewModule),
+          loadChildren: () => import('@alfresco/aca-content/preview').then((m) => m.PreviewModule),
           data: {
             navigateSource: 'favorites'
           }
@@ -398,7 +398,7 @@ export const CONTENT_LAYOUT_ROUTES: Route = {
         // deprecated, backwards compatibility with ACA 1.8
         {
           path: 'preview/:nodeId',
-          loadChildren: () => import('@alfresco/aca-preview').then((m) => m.PreviewModule),
+          loadChildren: () => import('@alfresco/aca-content/preview').then((m) => m.PreviewModule),
           data: {
             navigateSource: 'recent-files'
           }
@@ -447,7 +447,7 @@ export const CONTENT_LAYOUT_ROUTES: Route = {
         // deprecated, backwards compatibility with ACA 1.8
         {
           path: 'preview/:nodeId',
-          loadChildren: () => import('@alfresco/aca-preview').then((m) => m.PreviewModule),
+          loadChildren: () => import('@alfresco/aca-content/preview').then((m) => m.PreviewModule),
           data: {
             navigateSource: 'shared'
           }
@@ -508,7 +508,7 @@ export const CONTENT_LAYOUT_ROUTES: Route = {
         // deprecated, backwards compatibility with ACA 1.8
         {
           path: 'preview/:nodeId',
-          loadChildren: () => import('@alfresco/aca-preview').then((m) => m.PreviewModule),
+          loadChildren: () => import('@alfresco/aca-content/preview').then((m) => m.PreviewModule),
           data: {
             navigateSource: 'search'
           }
