--- projects/aca-preview/README.md
@@ -1,24 +1,0 @@
-# AcaPreview
-
-This library was generated with [Angular CLI](https://github.com/angular/angular-cli) version 14.1.0.
-
-## Code scaffolding
-
-Run `nx generate component component-name --project aca-preview` to generate a new component. You can also use `nx generate directive|pipe|service|class|guard|interface|enum|module --project aca-preview`.
-> Note: Don't forget to add `--project aca-preview` or else it will be added to the default project in your `angular.json` file. 
-
-## Build
-
-Run `nx build aca-preview` to build the project. The build artifacts will be stored in the `dist/` directory.
-
-## Publishing
-
-After building your library with `nx build aca-preview`, go to the dist folder `cd dist/aca-preview` and run `npm publish`.
-
-## Running unit tests
-
-Run `nx test aca-preview` to execute the unit tests via [Karma](https://karma-runner.github.io).
-
-## Further help
-
-To get more help on the NX use `nx help` or go check out the [NX documentation](https://nx.dev/getting-started/intro) page.
