--- projects/aca-preview/ng-package.json
@@ -1,7 +1,0 @@
-{
-  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/@alfresco/aca-preview",
-  "lib": {
-    "entryFile": "src/public-api.ts"
-  }
-}
