--- projects/aca-preview/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@alfresco/aca-preview",
-  "version": "0.0.1",
-  "license": "LGPL-3.0",
-  "peerDependencies": {
-    "@angular/common": "^14.1.0",
-    "@angular/core": "^14.1.0"
-  },
-  "dependencies": {
-    "tslib": "^2.3.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Alfresco/alfresco-content-app.git"
-  }
-}
