--- projects/aca-preview/project.json
@@ -1,56 +1,0 @@
-{
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "name": "aca-preview",
-  "projectType": "library",
-  "sourceRoot": "projects/aca-preview/src",
-  "prefix": "lib",
-  "targets": {
-    "build": {
-      "executor": "@angular-devkit/build-angular:ng-packagr",
-      "options": {
-        "project": "projects/aca-preview/ng-package.json"
-      },
-      "configurations": {
-        "production": {
-          "tsConfig": "projects/aca-preview/tsconfig.lib.prod.json"
-        },
-        "development": {
-          "tsConfig": "projects/aca-preview/tsconfig.lib.json"
-        }
-      },
-      "defaultConfiguration": "production",
-      "outputs": [
-        "{workspaceRoot}/dist/@alfresco/aca-preview"
-      ]
-    },
-    "test": {
-      "executor": "@angular-devkit/build-angular:karma",
-      "options": {
-        "codeCoverage": true,
-        "main": "projects/aca-preview/src/test.ts",
-        "tsConfig": "projects/aca-preview/tsconfig.spec.json",
-        "karmaConfig": "projects/aca-preview/karma.conf.js"
-      },
-      "configurations": {
-        "adfprod": {
-          "tsConfig": "projects/aca-preview/tsconfig.spec.adf.json"
-        }
-      }
-    },
-    "lint": {
-      "executor": "@angular-eslint/builder:lint",
-      "options": {
-        "lintFilePatterns": [
-          "projects/aca-preview/**/*.ts",
-          "projects/aca-preview/**/*.html"
-        ],
-        "cache": true,
-        "cacheLocation": ".eslintcache",
-        "ignorePath": ".eslintignore"
-      },
-      "outputs": [
-        "{options.outputFile}"
-      ]
-    }
-  }
-}
