--- projects/aca-preview/tsconfig.lib.prod.json
@@ -1,14 +1,0 @@
-/* To learn more about this file see: https://angular.io/config/tsconfig. */
-{
-  "extends": "./tsconfig.lib.json",
-  "compilerOptions": {
-    "declarationMap": false,
-    "paths": {
-      "@alfresco/aca-shared": ["dist/@alfresco/aca-shared"],
-      "@alfresco/aca-shared/store": ["dist/@alfresco/aca-shared/store"],
-    }
-  },
-  "angularCompilerOptions": {
-    "compilationMode": "partial"
-  }
-}
