--- projects/aca-preview/tsconfig.spec.adf.json
@@ -1,16 +1,0 @@
-{
-  "extends": "../../tsconfig.adf.json",
-  "compilerOptions": {
-    "outDir": "../../out-tsc/spec",
-    "types": [
-      "jasmine"
-    ]
-  },
-  "files": [
-    "src/test.ts"
-  ],
-  "include": [
-    "**/*.spec.ts",
-    "**/*.d.ts"
-  ]
-}
