+++ tsconfig.adf.json
@@ -42,7 +42,7 @@
       "@alfresco/aca-content/folder-rules": ["projects/aca-content/folder-rules/src/public-api.ts"],
       "@alfresco/aca-content": ["projects/aca-content/src/public-api.ts"],
       "@alfresco/aca-content/viewer": ["projects/aca-content/viewer/src/public-api.ts"],
-      "@alfresco/aca-preview": ["projects/aca-preview/src/public-api.ts"],
+      "@alfresco/aca-content/preview": ["projects/aca-content/preview/src/public-api.ts"],
       "package.json": ["package.json"],
       "*": ["./node_modules/*"]
     }
