+++ tsconfig.json
@@ -30,7 +30,7 @@
       "@alfresco/aca-content": ["projects/aca-content/src/public-api.ts"],
       "@alfresco/aca-content/about": ["projects/aca-content/about/src/public-api.ts"],
       "@alfresco/aca-content/viewer": ["projects/aca-content/viewer/src/public-api.ts"],
-      "@alfresco/aca-preview": ["projects/aca-preview/src/public-api.ts"],
+      "@alfresco/aca-content/preview": ["projects/aca-content/preview/src/public-api.ts"],
       "@alfresco/playwright-shared": ["projects/aca-playwright-shared/src/index.ts"],
       "package.json": ["package.json"]
     }
